backend/requirements.txt
fastapi
uvicorn[standard]
openai>=1.0.0
python-dotenv
requests

